+++ aggsender/db/aggsender_db_storage.go
@@ -373,13 +373,7 @@ func (a *AggSenderSQLStorage) GetLastSentCertificateHeaderWithProofIfInError(
 // and to allow for debugging and analysis of why they were not accepted.
 func (a *AggSenderSQLStorage) SaveNonAcceptedCertificate(
 	ctx context.Context, nonAcceptedCert *NonAcceptedCertificate) error {
-	return a.saveNonAcceptedCertificate(ctx, a.db, nonAcceptedCert)
-}
-
-func (a *AggSenderSQLStorage) saveNonAcceptedCertificate(
-	ctx context.Context, database dbtypes.DBer,
-	nonAcceptedCert *NonAcceptedCertificate) error {
-	tx, err := newTxer(ctx, database)
+	tx, err := newTxer(ctx, a.db)
 	if err != nil {
 		return fmt.Errorf("failed to create db transaction for non-accepted certificate persistence: %w", err)
 	}
